sixel: update the renderer and add new csi sequences
The updated renderer splits sixel images into sub-images whose width is
the same as the width of the sixel image and the height is the same as
the height of the text cell. This simplifies the implementation and
gives us many advantages. For example:
- Images now work properly with scrolling regions.
- The auto-delete feature is now enabled on the primary screen as well.
Previously, it could not be used because it caused problems with text
reflow.
- It's no longer necessary to cut out a large part of the image when
text is written in the middle of the image. We can just throw away one
sub-image.

The new added CSI sequences:
- CSI ? 80 h : Enable Sixel Display Mode (DECSDM). It places the image
  in the upper left corner, ignores the scrolling region (no scrolling)
  and does not change the cursor position.
- CSI ? 80 l : Disable Sixel Display Mode (DECSDM). (Default)
- CSI ? 80 $ p : Inquire if Sixel Display Mode is enabled. (DECRQM)
- CSI ? 8 4 5 2 h: Leave the cursor on the last row and on the right
  side of the image.
- CSI ? 8 4 5 2 l: Leave the cursor on the last row and the first
  column of the image. (Default)