Fix the behavior of tlinelen() on the alt screen
The tlinelen() function is using the ATTR_SET attribute to find the end
of a text line. That works fine on the main screen but not always on the
alt screen because an application might use spaces to clear the screen
and in that case it's impossible to find the end of the line when every
cell has the ATTR_SET attribute.

So we're now using a different algorithm on the alt screen, but we're
keeping the old algorithm on the main screen because it works more
accurately there.