Fix line wrapping when last character is wideglyph
If the cursor is on the last text cell of the line and we try to print
there a wideglyph that takes up two cells, we have to move it to the
next line. In this special case, the ATTR_WRAP attribute must be set to
the penultimate cell of the current line because there is now an extra
space at the end of the line. This way the lines are joined correctly
when the text is reflowed. The issue arises when something has already
been printed to the last cell, because the tlinelen() and tiswrapped()
functions can no longer find the ATTR_WRAP attribute, which breaks text
reflow. So we need to clear the last cell in this special case to fix
the issue.