Avoid redrawing previous cursor line twice
The ligature and wideglyph patches require the previous cursor line to
be redrawn when the cursor is moved, which causes two issues:

1. When a cursor key is pressed down and released, the previous cursor
line is redrawn twice. This can be avoided by redrawing the line only
when the cursor actually moves or changes.

2. The xdrawcursor function that redraws the previous cursor line
doesn't know if the drawregion function has already done it, so we might
end up drawing it twice. This can be solved by calling xdrawcursor()
before we call drawregion(), and let xdrawcursor() draw the previous and
current cursor lines when needed and mark them clean so that
drawregion() doesn't redraw them again.

Although redrawing the line twice is not a big issue, we don't want to
waste any CPU cycles on unnecessary work.