Add support for querying the cursor style
The Yazi file manager uses the DECRQSS sequence to query the cursor
style on startup so that it can restore the cursor on exit. Since st
does not support the DECRQSS sequence, this patch adds partial support
for it. For now, only the cursor style (DECSCUSR) can be requested via
DECRQSS.