openurlonclick: copy urls to clipboard with right click
After this patch, you can copy the detected urls and hyperlinks to the
clipboard by pressing the right mouse button. The file protocol and
hostname will be removed from local files, so that when you paste the
file paths into the terminal, you don't have to remove the protocol and
hostname manually.