openurlonclick: change modkey to behave like forcemousemod
This update changes url_opener_modkey to behave like forcemousemod when
the terminal is in mouse mode. So if url_opener_modkey is set to the
control key, you no longer need to press both the control and
forcemousemod keys when clicking links. You only need to press the
control key, which makes the modkey more consistent to use in all modes.
However, if the modkey is set to XK_ANY_MOD, you still need to press the
forcemousemod key in mouse mode.