Move XftDrawSetClip calls out of xdrawglyphfontspecs
When st draws a line, it divides it into smaller line segments with the
same attributes and colors. Then it sets a clipping region to each
segment and draws them. Unfortunately, the XftDrawSetClip function calls
break the ability of the XRender and Xlib libraries to batch the
FillRectangle requests together, which increases the xorg usage and
degrades performance. In worst case scenarios, where there is a lot of
different colored text on the screen, the rendering speed can drop from
200 fps to 45 fps or even lower. This happens in the vanilla st as well.

So this fix moves the XftDrawSetClip calls out of the
xdrawglyphfontspecs function and only sets the clipping region once per
line in the xdrawline function. This change fixes the performance drops
in those worst case scenarios. However, this cannot save the extreme
cases where each letter has different attributes or colors, because in
those cases the huge number of the XftDrawRect and XftDrawGlyphFontSpec
calls will drop the performance anyway. Fortunately, these extreme cases
are not normal cases.

It's important to mention that this fix won't help much without the wide
glyph patch, because the patch changes the drawing order, which allows
the FillRectangle requests to be batched together for better
performance.