csi: check for private marker in 'u' case
The kitty terminal has added the new sequence "CSI ? u" for querying the
keyboard enhancement flags. Since st does not check the private marker
in that sequence, it interprets the sequence as the restore cursor
sequence "CSI u" and may move the cursor to the wrong place when an
application queries for the enhancement flags. To fix the issue, this
patch adds a check for the private marker.

Fixes #15