Fix ESC key behavior
When you start flash mode directly from your shell and press ESC, it
should exit back to the shell. But if you start flash mode from move or
select mode and then press ESC, it should go back to the previous mode
and not exit to the shell. This fixes the latter case.