Optimize csireset()
When st begins to process an escape sequence, it calls the csireset
function that reset the entire csiescseq structure. The size of
csiescseq is 600 bytes in st and 984 bytes in st-sx, so it is a fairly
large structure to be reset for each escape sequence. So I investigated
if this could be improved and I found that we only need to reset the
fields len, priv and arg in csiescseq. Which means that only 73 out of
984 bytes need to be reset and that is what this new optimized code
does. Although it may not make a noticeable difference in normal use, it
still reduces latency when processing escape sequences.