keyboard select: fix label placement issue in flash mode
If a line ends with the word being searched for, the flash label is
placed on the last character of the word, even if there is space for it
after the word. For example, if a line ends with the word "flash" and
you type "flash", the label will be placed on the letter h regardless of
whether there is space for it after the word or not. This patch aims to
fix this issue and put the labels where they should be.