Fix another selection clearing bug
When you select text from the command line and then insert or delete
letters before the selection, the text moves right or left, but the
selection remains in place. This isn't correct behavior, because the
selection should either move with the text or be cleared. Since other
terminals simply clear the selection in these cases, this fix follows
the same practice.

N.B. This bug is also present in the vanilla st.