Add the ability to disable unfocused bg opacity
This commit added the ability to disable the unfocused bg opacity
by setting alphaUnfocused to -1. Doing so will cause alpha to be
used for both the focused and unfocused bg opacity.

One can instead set alpha and alphaUnfocused to be the same value,
however these value will deviate as soon as changealpha is used.
Setting alphaUnfocused to -1 will ensure the focused and unfocused
bg opacity will always be the same; taking into account changealpha.