Support multiple parameters in OSC 4 and 104 sequences
The specification says that OSC 4 and 104 sequences can have multiple
color parameters, but the current implementation is only able to set or
report a single color. This patch fixes this issue.

Excerpt from the spec:

  Ps = 4 ; c ; spec =>  Change Color Number c to the color specified by
  spec. The spec can be a name or RGB specification as per XParseColor.
  Any number of c/spec pairs may be given.

  Ps = 1 0 4 ; c =>  Reset Color Number c.  It is reset to the color
  specified by the corresponding X resource.  Any number of c parameters
  may be given.

  https://invisible-island.net/xterm/ctlseqs/ctlseqs.html#h3-Operating-System-Commands

With this patch, Microsoft Edit is able to use the terminal color theme
instead of the default color theme. It's the first application I've seen
that uses multiple color parameters in OSC 4 sequences.