xresources: remove the FocusOut sequence
The Xresources patch sends the FocusOut sequence to the running app when
xresources is reloaded. Based on the code comment, the purpose of the
sequence is to trigger the app to redraw its content, although it does
not specify which apps require this. The problem with the sequence is
that some apps don't know how to handle them correctly. For example, the
nano editor incorrectly thinks that the user pressed Ctrl+O when it
receives the FocusOut sequence. So for this reason, this update removes
the sequence until we know where and when it is needed.