Correct color register count
st-sx advertises 1024 color registers in XTSMGRAPHICS, but in fact only 1023
were available because 0 is used as a pseudo-register for background (or the
lack of it).  This is a problem if a program wishes to actually use all 1024
registers, so I've added an extra slot for the pseudo-register.

For example, following ought to print a red bar, but was printing a green one:

printf '\033P;0q#1022;2;99;0;0#1023;2;0;99;0;0#1022!100~-\033\\\n'