Fix DECRC to restore cursor wrap state
st does not restore the cursor wrap state when it executes the DECRC
control sequence. It seems like it's a bug because the standard (see
https://github.com/mattiase/wraptest) states that the wrap state should
be restored, and all terminal emulators I have tested restore it except
Konsole. This update ensures that all cursor states, including the wrap
state, are restored with DECRC.

Test case that should print the numbers from 1 to 6:
printf "\e[$(($COLUMNS-2))G123\e7\e8456\n"