Fix some sequences to reset cursor wrap state
According to the DEC STD 070 manual and wraptest, the ED, EL, ECH, DCH,
ICH, IND, RI, and DECAWM escape sequences should reset the last column
flag (LCF), but st does not do that. It's not a big issue, since many
other terminal emulators also don't adhere to the standard. However,
since XTerm does and is regarded as the reference implementation, we
should follow its behavior.

So, this update fixes the above sequences to reset the LFC. However, I
have decided to deviate from the standard regarding EL0, so that now it
does nothing if LCF is set. The reason is that GNU grep sometimes emits
EL0 when LCF is active, which can cause the last character of the line
to be dropped. VTE terminals use the same solution to fix this issue.

References:
https://github.com/mattiase/wraptest
https://bugzilla.gnome.org/show_bug.cgi?id=740789